+++ src/database/factories/CommentFactory.php
@@ -2,6 +2,7 @@
 
 namespace Database\Factories;
 
+use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
@@ -17,8 +18,7 @@ class CommentFactory extends Factory
     public function definition()
     {
         return [
-            'user_id' => 1,
-            'post_id' => 1,
+            'user_id' => User::inRandomOrder()->first()->id,
             'comment' => $this->faker->sentence(),
         ];
     }