+++ src/database/factories/PostFactory.php
@@ -2,6 +2,8 @@
 
 namespace Database\Factories;
 
+use App\Models\Category;
+use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
@@ -24,8 +26,8 @@ class PostFactory extends Factory
             'title' => $title,
             'slug' => $slug,
             'keywords' => $this->faker->word(),
-            'category_id' => '1',
-            'user_id' => 1,
+            'category_id' => Category::inRandomOrder()->first()->id,
+            'user_id' => User::inRandomOrder()->first()->id,
             'post' => $this->faker->paragraph(),
             'banner_image_url' => null,
             'published' => 1,