+++ src/database/migrations/2020_01_27_213243_insert_admin_account.php
@@ -1,11 +1,10 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\Schema;
-use App\Models\User;
-use App\Models\Setting;
 use App\Models\Category;
+use App\Models\Setting;
+use App\Models\User;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Str;
 
 class InsertAdminAccount extends Migration
 {