+++ src/database/seeders/PostsTableSeeder.php
@@ -2,8 +2,8 @@
 
 namespace Database\Seeders;
 
-use Illuminate\Database\Seeder;
 use App\Models\Post;
+use Illuminate\Database\Seeder;
 
 class PostsTableSeeder extends Seeder
 {
@@ -14,6 +14,9 @@ class PostsTableSeeder extends Seeder
      */
     public function run()
     {
-        Post::factory()->count(5)->create();
+        Post::factory()
+            ->hasComments(5)
+            ->count(5)
+            ->create();
     }
 }